pyproject.toml
[project]
name = "frankasset"
version = "0.1.0"
description = "Agentic AI backend for brand classification using Google ADK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "structlog>=23.0.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "google-cloud-aiplatform>=1.38.0",
    "google-generativeai>=0.3.0",
    "google-adk>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
